+++ photutils/__init__.py
@@ -47,14 +47,9 @@ __depr__[detection] = ('StarFinderBase', 'DAOStarFinder', 'IRAFStarFinder',
                        'find_peaks', 'StarFinder')
 
 __depr__[psf] = ('EPSFFitter', 'EPSFBuilder', 'EPSFStar', 'EPSFStars',
-                 'LinkedEPSFStar', 'extract_stars', 'DAOGroup', 'DBSCANGroup',
-                 'GroupStarsBase', 'NonNormalizable', 'FittableImageModel',
+                 'LinkedEPSFStar', 'extract_stars', 'FittableImageModel',
                  'EPSFModel', 'GriddedPSFModel', 'IntegratedGaussianPRF',
-                 'PRFAdapter', 'BasicPSFPhotometry',
-                 'IterativelySubtractedPSFPhotometry', 'DAOPhotPSFPhotometry',
-                 'prepare_psf_model',
-                 'get_grouped_psf_model', 'subtract_psf',
-                 'resize_psf', 'create_matching_kernel',
+                 'PRFAdapter', 'resize_psf', 'create_matching_kernel',
                  'SplitCosineBellWindow', 'HanningWindow', 'TukeyWindow',
                  'CosineBellWindow', 'TopHatWindow')