+++ photutils/psf/tests/test_photometry.py
@@ -10,8 +10,7 @@ from astropy.modeling.fitting import LMLSQFitter, SimplexLSQFitter
 from astropy.modeling.models import Gaussian1D, Gaussian2D, custom_model
 from astropy.nddata import NDData, StdDevUncertainty
 from astropy.table import QTable, Table
-from astropy.utils.exceptions import (AstropyDeprecationWarning,
-                                      AstropyUserWarning)
+from astropy.utils.exceptions import AstropyUserWarning
 from numpy.testing import assert_allclose, assert_equal
 
 from photutils.background import LocalBackground, MMMBackground
@@ -20,7 +19,6 @@ from photutils.datasets import (make_model_image, make_noise_image,
 from photutils.detection import DAOStarFinder
 from photutils.psf import (IntegratedGaussianPRF, IterativePSFPhotometry,
                            PSFPhotometry, SourceGrouper, make_psf_model)
-from photutils.psf.photometry_depr import DAOGroup
 from photutils.utils._optional_deps import HAS_SCIPY
 from photutils.utils.cutouts import _overlap_slices as overlap_slices
 from photutils.utils.exceptions import NoDetectionsWarning
@@ -85,16 +83,11 @@ def test_invalid_inputs():
         with pytest.raises(ValueError, match=match):
             _ = PSFPhotometry(model, shape)
 
-    kwargs = {'grouper': 1, 'finder': 1, 'fitter': 1}
+    kwargs = {'finder': 1, 'fitter': 1}
     for key, val in kwargs.items():
         match = f"'{key}' must be a callable object"
         with pytest.raises(TypeError, match=match):
             _ = PSFPhotometry(model, 1, **{key: val})
-
-    match = 'Invalid grouper class. Please use SourceGrouper.'
-    with pytest.raises(ValueError, match=match):
-        with pytest.warns(AstropyDeprecationWarning):
-            _ = PSFPhotometry(model, 1, grouper=DAOGroup(1))
 
     match = 'localbkg_estimator must be a LocalBackground instance'
     with pytest.raises(ValueError, match=match):
@@ -299,18 +292,6 @@ def test_psf_photometry_compound_psfmodel(test_data):
     phot = psfphot(data, error=error)
     assert isinstance(phot, QTable)
     assert len(phot) == len(sources)
-
-    psf_model2 = psf_model.copy()
-    psf_model2.xname = psf_model2.x_name
-    psf_model2.yname = psf_model2.y_name
-    psf_model2.fluxname = psf_model2.flux_name
-    del psf_model2.x_name, psf_model2.y_name, psf_model2.flux_name
-
-    psfphot2 = PSFPhotometry(psf_model2, fit_shape, finder=finder,
-                             aperture_radius=4)
-    phot2 = psfphot2(data, error=error)
-    assert isinstance(phot2, QTable)
-    assert len(phot2) == len(sources)
 
     # test results when fit does not converge (fitter_maxiters=10)
     match = r'One or more fit\(s\) may not have converged.'