+++ photutils/psf/tests/test_utils.py
@@ -10,15 +10,12 @@ from astropy.modeling.fitting import LevMarLSQFitter
 from astropy.modeling.models import Const2D, Gaussian2D, Moffat2D
 from astropy.nddata import NDData
 from astropy.table import Table
-from astropy.utils.exceptions import AstropyDeprecationWarning
 from numpy.testing import assert_allclose, assert_equal
 
 from photutils import datasets
 from photutils.detection import find_peaks
-from photutils.psf import (BasicPSFPhotometry, DAOGroup, EPSFBuilder,
-                           IntegratedGaussianPRF, extract_stars,
-                           get_grouped_psf_model, grid_from_epsfs,
-                           make_psf_model, prepare_psf_model, subtract_psf)
+from photutils.psf import (EPSFBuilder, extract_stars, grid_from_epsfs,
+                           make_psf_model)
 from photutils.psf.utils import (_integrate_model, _interpolate_missing_data,
                                  _InverseShift)
 from photutils.utils._optional_deps import HAS_SCIPY
@@ -222,8 +219,6 @@ def test_make_psf_model_offset():
     """
     Test to ensure the offset is in the correct direction.
     """
-    from astropy.modeling.models import Moffat2D
-
     moffat = Moffat2D(x_0=0, y_0=0, alpha=4.8)
     psfmod1 = make_psf_model(moffat.copy(), x_name='x_0', y_name='y_0',
                              normalize=False)
@@ -233,180 +228,6 @@ def test_make_psf_model_offset():
     psfmod2.offset_0 = 10
 
     assert moffat(10, 0) == psfmod1(10, 0) == psfmod2(10, 0) == 1.0
-
-
-# we set the tolerances in flux to be 2-3% because the shape parameters of
-# the guessed version are known to be wrong.
-@pytest.mark.parametrize('prepkwargs, tols', [
-                         (dict(xname='x_0', yname='y_0', fluxname=None,
-                               renormalize_psf=True), (1e-3, 0.02)),
-                         (dict(xname=None, yname=None, fluxname=None,
-                               renormalize_psf=True), (1e-3, 0.02)),
-                         (dict(xname=None, yname=None, fluxname=None,
-                               renormalize_psf=False), (1e-3, 0.03)),
-                         (dict(xname='x_0', yname='y_0', fluxname='amplitude',
-                               renormalize_psf=False), (1e-3, None)),
-                         ])
-@pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-def test_prepare_psf_model(moffat_source, prepkwargs, tols):
-    """
-    Test that prepare_psf_model behaves as expected for fitting.
-    """
-    with pytest.warns(AstropyDeprecationWarning):
-        model, (xx, yy, data) = moffat_source
-        fitter = LevMarLSQFitter()
-
-        # a close-but-wrong "guessed Moffat"
-        guess_moffat = Moffat2D(x_0=.1, y_0=-.05, gamma=1.01,
-                                amplitude=model.amplitude * 1.01, alpha=4.79)
-        if prepkwargs['renormalize_psf']:
-            # definitely very wrong, so this ensures the re-normalization
-            # works
-            guess_moffat.amplitude = 5.0
-
-        if prepkwargs['xname'] is None:
-            guess_moffat.x_0 = 0
-        if prepkwargs['yname'] is None:
-            guess_moffat.y_0 = 0
-
-        psfmod = prepare_psf_model(guess_moffat, **prepkwargs)
-        xytol, fluxtol = tols
-
-        fit_psfmod = fitter(psfmod, xx, yy, data)
-
-        if xytol is not None:
-            assert np.abs(getattr(fit_psfmod, fit_psfmod.xname)) < xytol
-            assert np.abs(getattr(fit_psfmod, fit_psfmod.yname)) < xytol
-        if fluxtol is not None:
-            assert np.abs(1 - getattr(fit_psfmod, fit_psfmod.fluxname)) < fluxtol
-
-        # ensure the model parameters did not change
-        assert fit_psfmod.psfmodel.gamma == guess_moffat.gamma
-        assert fit_psfmod.psfmodel.alpha == guess_moffat.alpha
-        if prepkwargs['fluxname'] is None:
-            assert fit_psfmod.psfmodel.amplitude == guess_moffat.amplitude
-
-
-@pytest.mark.filterwarnings('ignore:aperture_radius is None and could not '
-                            'be determined by psf_model')
-@pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-def test_prepare_psf_model_offset():
-    """
-    Regression test to ensure the offset is in the correct direction.
-    """
-    with pytest.warns(AstropyDeprecationWarning):
-        norm = False
-        sigma = 3.0
-        amplitude = 1.0 / (2 * np.pi * sigma**2)
-        xcen = ycen = 0.0
-        psf0 = Gaussian2D(amplitude, xcen, ycen, sigma, sigma)
-        psf1 = prepare_psf_model(psf0, xname='x_mean', yname='y_mean',
-                                 renormalize_psf=norm)
-        psf2 = prepare_psf_model(psf0, renormalize_psf=norm)
-        psf3 = prepare_psf_model(psf0, xname='x_mean', renormalize_psf=norm)
-        psf4 = prepare_psf_model(psf0, yname='y_mean', renormalize_psf=norm)
-
-        yy, xx = np.mgrid[0:101, 0:101]
-        psf = psf1.copy()
-        xval = 48
-        yval = 52
-        flux = 14.51
-        psf.x_mean_2 = xval
-        psf.y_mean_2 = yval
-        data = psf(xx, yy) * flux
-
-        group_maker = DAOGroup(2)
-        bkg_estimator = None
-        fitshape = 7
-        init_guesses = Table([[46.1], [57.3], [7.1]],
-                             names=['x_0', 'y_0', 'flux_0'])
-
-        phot1 = BasicPSFPhotometry(group_maker=group_maker,
-                                   bkg_estimator=bkg_estimator,
-                                   fitshape=fitshape, psf_model=psf1)
-        tbl1 = phot1(image=data, init_guesses=init_guesses)
-
-        phot2 = BasicPSFPhotometry(group_maker=group_maker,
-                                   bkg_estimator=bkg_estimator,
-                                   fitshape=fitshape, psf_model=psf2)
-        tbl2 = phot2(image=data, init_guesses=init_guesses)
-
-        phot3 = BasicPSFPhotometry(group_maker=group_maker,
-                                   bkg_estimator=bkg_estimator,
-                                   fitshape=fitshape, psf_model=psf3)
-        tbl3 = phot3(image=data, init_guesses=init_guesses)
-
-        phot4 = BasicPSFPhotometry(group_maker=group_maker,
-                                   bkg_estimator=bkg_estimator,
-                                   fitshape=fitshape, psf_model=psf4)
-        tbl4 = phot4(image=data, init_guesses=init_guesses)
-
-        assert_allclose((tbl1['x_fit'][0], tbl1['y_fit'][0],
-                         tbl1['flux_fit'][0]), (xval, yval, flux))
-        assert_allclose((tbl2['x_fit'][0], tbl2['y_fit'][0],
-                         tbl2['flux_fit'][0]), (xval, yval, flux))
-        assert_allclose((tbl3['x_fit'][0], tbl3['y_fit'][0],
-                         tbl3['flux_fit'][0]), (xval, yval, flux))
-        assert_allclose((tbl4['x_fit'][0], tbl4['y_fit'][0],
-                         tbl4['flux_fit'][0]), (xval, yval, flux))
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-def test_get_grouped_psf_model():
-    igp = IntegratedGaussianPRF(sigma=1.2)
-    tab = Table(names=['x_0', 'y_0', 'flux_0'],
-                data=[[1, 2], [3, 4], [0.5, 1]])
-    pars_to_set = {'x_0': 'x_0', 'y_0': 'y_0', 'flux_0': 'flux'}
-
-    with pytest.warns(AstropyDeprecationWarning):
-        gpsf = get_grouped_psf_model(igp, tab, pars_to_set)
-
-    assert gpsf.x_0_0 == 1
-    assert gpsf.y_0_1 == 4
-    assert gpsf.flux_0 == 0.5
-    assert gpsf.flux_1 == 1
-    assert gpsf.sigma_0 == gpsf.sigma_1 == 1.2
-
-
-@pytest.fixture(params=[0, 1, 2])
-def prf_model(request):
-    # use this instead of pytest.mark.parameterize as we use scipy and
-    # it still calls that even if not HAS_SCIPY is set...
-    with pytest.warns(AstropyDeprecationWarning):
-        prfs = [IntegratedGaussianPRF(sigma=1.2),
-                Gaussian2D(x_stddev=2),
-                prepare_psf_model(Gaussian2D(x_stddev=2),
-                                  renormalize_psf=False)]
-    return prfs[request.param]
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-def test_get_grouped_psf_model_submodel_names(prf_model):
-    """
-    Verify that submodel tagging works.
-    """
-    tab = Table(names=['x_0', 'y_0', 'flux_0'],
-                data=[[1, 2], [3, 4], [0.5, 1]])
-    pars_to_set = {'x_0': 'x_0', 'y_0': 'y_0', 'flux_0': 'flux'}
-
-    with pytest.warns(AstropyDeprecationWarning):
-        gpsf = get_grouped_psf_model(prf_model, tab, pars_to_set)
-        # There should be two submodels one named 0 and one named 1
-        assert len([submodel for submodel in gpsf.traverse_postorder()
-                    if submodel.name == 0]) == 1
-        assert len([submodel for submodel in gpsf.traverse_postorder()
-                    if submodel.name == 1]) == 1
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason='scipy is required')
-def test_subtract_psf():
-    with pytest.warns(AstropyDeprecationWarning):
-        psf = IntegratedGaussianPRF(sigma=1.0)
-        posflux = SOURCES.copy()
-        for n in posflux.colnames:
-            posflux.rename_column(n, n.split('_')[0] + '_fit')
-        residuals = subtract_psf(image, psf, posflux)
-        assert np.max(np.abs(residuals)) < 0.0052
 
 
 @pytest.mark.remote_data